refactor: extract TableField subcomponents to own files
