+++ frontend/src/templates/Field/Table/TableField.tsx
@@ -1,52 +1,25 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
 /**
  * @precondition Must have a parent `react-hook-form#FormProvider` component.
  */
 import { useCallback, useMemo } from 'react'
-import {
-  Controller,
-  ControllerRenderProps,
-  FieldError,
-  FormProvider,
-  useFieldArray,
-  useForm,
-  useFormContext,
-} from 'react-hook-form'
-import { BiPlus } from 'react-icons/bi'
-import { useTable, UseTableCellProps } from 'react-table'
-import {
-  FormControl,
-  Stack,
-  Table,
-  Tbody,
-  Td,
-  Text,
-  Th,
-  Thead,
-  Tr,
-} from '@chakra-ui/react'
-import { get, times } from 'lodash'
+import { FormProvider, useFieldArray, useForm } from 'react-hook-form'
+import { useTable } from 'react-table'
+import { Table, Tbody, Td, Th, Thead, Tr } from '@chakra-ui/react'
+import { times } from 'lodash'
 import { Merge, RequireAllOrNone } from 'type-fest'
 
-import {
-  BasicField,
-  Column,
-  FormFieldWithId,
-  ShortTextColumnBase,
-  TableFieldBase,
-} from '~shared/types/field'
+import { Column, FormFieldWithId, TableFieldBase } from '~shared/types/field'
 
-import { createShortTextValidationRules } from '~utils/fieldValidation'
-import Button from '~components/Button'
 import { FormErrorMessage } from '~components/FormControl/FormErrorMessage/FormErrorMessage'
-import FormLabel from '~components/FormControl/FormLabel'
-import Input from '~components/Input'
 
 import { BaseFieldProps } from '../FieldContainer'
 
+import { AddRowFooter } from './AddRowFooter'
+import { ColumnCell } from './ColumnCell'
+import { ColumnHeader } from './ColumnHeader'
 import { TableFieldContainer } from './TableFieldContainer'
 
-type ColumnWithId<T = Column> = T & { _id: string }
+export type ColumnWithId<T = Column> = T & { _id: string }
 export type TableFieldSchema = Merge<
   RequireAllOrNone<
     FormFieldWithId<TableFieldBase>,
@@ -58,77 +31,13 @@ export interface TableFieldProps extends BaseFieldProps {
   schema: TableFieldSchema
 }
 
-const ShortTextColumnCell = ({
-  schema,
-  inputName,
-}: {
-  schema: ColumnWithId<ShortTextColumnBase>
-  inputName: string
-}) => {
-  const rules = useMemo(() => createShortTextValidationRules(schema), [schema])
-
-  return (
-    <Controller
-      name={inputName}
-      rules={rules}
-      render={({ field }) => <Input {...field} />}
-    />
-  )
-}
-
-/**
- * Renderer for each column cell in the table schema.
- */
-const ColumnCell = ({
-  schemaId,
-  row,
-  column,
-  columnSchema,
-}: UseTableCellProps<Record<string, unknown>, string> & {
-  schemaId: string
-  columnSchema: ColumnWithId
-}) => {
-  const {
-    formState: { errors },
-  } = useFormContext()
-
-  const inputName = useMemo(
-    () => `${schemaId}.${row.index}.${column.id}`,
-    [column.id, row.index, schemaId],
-  )
-
-  const cellError: FieldError | undefined = get(errors, inputName)
-
-  const renderedColumnCell = useMemo(() => {
-    switch (columnSchema.columnType) {
-      case BasicField.ShortText:
-        return (
-          <ShortTextColumnCell schema={columnSchema} inputName={inputName} />
-        )
-      default:
-        return null
-    }
-  }, [columnSchema, inputName])
-
-  return (
-    <FormControl
-      isRequired={columnSchema.required}
-      isInvalid={!!cellError}
-      mb={6}
-    >
-      {renderedColumnCell}
-      <FormErrorMessage>{cellError?.message}</FormErrorMessage>
-    </FormControl>
-  )
-}
-
 export const TableField = ({
   schema,
   questionNumber,
 }: TableFieldProps): JSX.Element => {
   const columnsData = useMemo(() => {
     return schema.columns.map((c) => ({
-      Header: c.title,
+      Header: <ColumnHeader title={c.title} isRequired={c.required} />,
       accessor: c._id,
       Cell: ColumnCell,
     }))
@@ -156,7 +65,6 @@ export const TableField = ({
     control: formMethods.control,
     name: schema._id,
   })
-
   const { getTableProps, getTableBodyProps, headerGroups, rows, prepareRow } =
     useTable({ columns: columnsData, data: fields })
 
@@ -174,12 +82,17 @@ export const TableField = ({
         )}
       >
         <TableFieldContainer schema={schema} questionNumber={questionNumber}>
-          <Table {...getTableProps()}>
+          <Table {...getTableProps()} d="inline-block" overflowX="auto">
             <Thead>
               {headerGroups.map((headerGroup) => (
                 <Tr {...headerGroup.getHeaderGroupProps()}>
-                  {headerGroup.headers.map((column) => (
-                    <Th {...column.getHeaderProps()}>
+                  {headerGroup.headers.map((column, _idx, array) => (
+                    <Th
+                      {...column.getHeaderProps()}
+                      w={`calc(100%/${array.length})`}
+                      minW="12rem"
+                      id={column.id}
+                    >
                       {column.render('Header')}
                     </Th>
                   ))}
@@ -204,6 +117,10 @@ export const TableField = ({
               })}
             </Tbody>
           </Table>
+          {/* Assume any error is required, since that's the only error (for now) */}
+          <FormErrorMessage my="0.75rem">
+            Please fill in the required fields
+          </FormErrorMessage>
         </TableFieldContainer>
         {schema.addMoreRows && (
           <AddRowFooter
@@ -215,33 +132,5 @@ export const TableField = ({
         <button type="submit">Submit</button>
       </form>
     </FormProvider>
-  )
-}
-
-interface AddRowFooterProps {
-  handleAddRow: () => void
-  currentRows: number
-  maxRows: number
-}
-const AddRowFooter = ({
-  currentRows,
-  maxRows,
-  handleAddRow,
-}: AddRowFooterProps) => {
-  return (
-    <Stack direction="row" justify="space-between" align="center">
-      <Button
-        isDisabled={currentRows >= maxRows}
-        leftIcon={<BiPlus fontSize="1.5rem" />}
-        type="button"
-        onClick={handleAddRow}
-      >
-        Add another row
-      </Button>
-
-      <Text textStyle="body-2" color="secondary.400">
-        {currentRows} out of max {maxRows} rows
-      </Text>
-    </Stack>
   )
 }
